refactor: moving files info exportProcessor
