+++ src/tab/exportProcessor/formatter/formatText.js
