+++ src/tab/exportProcessor/scraper/extractor/ExtractorMaxAIGoogle.js
